fix(scorecard): fold NFS into plugin, drop legacy re-export, leave alpha for translations
