+++ workspaces/scorecard/plugins/scorecard/src/blueprints/index.ts
